requirements.txt
Flask~=3.1.0
SQLAlchemy~=1.4.46
WTForms~=3.2.1
Werkzeug~=3.1.3
click~=8.1.8
nltk~=3.9.1
beautifulsoup4~=4.13.3
Jinja2==3.1.6
email-validator==2.2.0
genanki==0.13.1
requests~=2.32.3
Pillow==11.1.0
babel==2.17.0
python-dotenv==1.1.0
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.7
flask-babel==4.0.0
Flask-WTF==1.2.2
psycopg2-binary==2.9.10
flask-mail==0.10.0
itsdangerous==2.2.0
ebooklib==0.18.0        # Для работы с EPUB файлами
pypdf>=4.0.0            # Для работы с PDF файлами
python-docx==1.1.2      # Для работы с DOCX файлами
lxml==5.3.1             # Требуется для ebooklib и beautifulsoup4
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
setuptools==78.1.0
psutil==7.0.0
Flask-JWT-Extended==4.7.1
bleach==6.2.0           # Для санитизации HTML контента
marshmallow==3.23.2     # Для валидации данных
Flask-Limiter==3.12     # Для rate limiting
celery==5.4.0           # Для фоновых задач
redis==5.2.1            # Message broker для Celery
APScheduler==3.10.4     # Задачи по расписанию (напоминания)
pytz==2024.1            # Часовые пояса для Telegram бота
Flask-Compress==1.15    # Gzip compression for HTTP responses
python-json-logger==2.0.7  # JSON-formatted structured logging